Refactor redis and kafka interactions
