+++ pkg/aggregator/file/types.go
@@ -1,8 +1,8 @@
 package file
 
 import (
-	"github.com/gargath/pleiades/pkg/aggregator"
 	"github.com/gargath/pleiades/pkg/spinner"
+	"github.com/gargath/pleiades/pkg/util"
 	"github.com/go-redis/redis/v8"
 )
 
@@ -10,7 +10,7 @@ import (
 type Aggregator struct {
 	File    *Opts
 	stop    chan (bool)
-	Redis   *aggregator.RedisOpts
+	Redis   *util.RedisOpts
 	r       *redis.Client
 	spinner *spinner.Spinner
 }
