+++ pkg/aggregator/kafka/types.go
@@ -1,8 +1,8 @@
 package kafka
 
 import (
-	"github.com/gargath/pleiades/pkg/aggregator"
 	"github.com/gargath/pleiades/pkg/spinner"
+	"github.com/gargath/pleiades/pkg/util"
 	"github.com/go-redis/redis/v8"
 	"github.com/segmentio/kafka-go"
 )
@@ -11,7 +11,7 @@ import (
 type Aggregator struct {
 	Kafka   *Opts
 	stop    chan (bool)
-	Redis   *aggregator.RedisOpts
+	Redis   *util.RedisOpts
 	r       *redis.Client
 	k       *kafka.Reader
 	spinner *spinner.Spinner
