+++ pkg/util/types.go
@@ -1,0 +1,6 @@
+package util
+
+// RedisOpts contains redis configuration
+type RedisOpts struct {
+	RedisAddr string
+}
